Added WYSISYG View for the gherkin elemet of the software + start of GUI implementation
After trying to work with the JS version of the development, it has been
deemed as too difficult within the time frame involved. Therefore a move
has been made to switch back to the older Java version in this
repository. Changes include:
Addition of WYSIWYG View for the Gherkin element of the GUI
Addition of sample Cucumber + Step Def scripts
Start of creation for actual code of JavaFX GUI